Clean up module structure.
Consolidate error classes.

Fixes #10.
Fixes #11.
