--- pdbx/writer/__init__.py
@@ -1,11 +1,0 @@
-"""Python PDBx/mmCIF writer resources.
-
-All of the code in this repository is based on http://mmcif.wwpdb.org/.
-Specifically, this code is directly derived from
-http://mmcif.wwpdb.org/docs/sw-examples/python/src/pdbx.tar.gz linked from
-http://mmcif.wwpdb.org/docs/sw-examples/python/html/.
-
-See http://mmcif.wwpdb.org/docs/sw-examples/python/html/ for more information
-about this package, including examples.
-"""
-__all__ = ["writer"]
